third_party/blink/web_tests/external/wpt/scroll-animations/css/scroll-timeline-dynamic.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Switching timelines and pausing at the same time [immediate]
  assert_equals: expected "120px" but got "100px"
[FAIL] Switching timelines and pausing at the same time [scroll]
  assert_equals: expected "120px" but got "100px"
Harness: the test ran to completion.
